+++ workspace/Futbol5/src/ordenamiento/UltimasCalificaciones.java
@@ -12,7 +12,7 @@ public class UltimasCalificaciones extends Criterio{
 		this.cantidad = cantidad;
 	}
 
-	public int valuarJugador(Jugador j) throws FutbolException {
+	public double valuarJugador(Jugador j) throws FutbolException {
 
 		int total = 0;
 		int size = j.calificaciones.size();
@@ -22,10 +22,10 @@ public class UltimasCalificaciones extends Criterio{
 			for(int i=0; i < cantidad; i++){
 				
 				Calificacion c = j.calificaciones.get(size-i-1);
-				total += c.getNota();				
+				total += c.getNota();		
 			}
-			
-			return total / cantidad;
+
+			return (double) total / cantidad;
 		}
 		else throw new FutbolException("No recibio calificaciones aun.");
 	}
